Please complete all the features that have not yet been implemented for example the edit strategy icon when pressed says not yet implemented. Can this be implemented such that when clicked I can view the python code version all form with all the paramenters that need configuring for the agent according to pydantic ai documentation as the agents are pydantic-ai based. After that implement the following : Agents By default the agents currently available in the system cannot be deleted but can be deactivated, however prompts can be updated during the configuration process by the user (in case use messes up hardcode default settings that can be loaded when clicked just be sure to confirm if user really want to load default agent settings)
Monitoring
When view running strategies is clicked, open the page and show charts of performance of the respective strategies and all the trades (with filters of up to 100 rows) that were opened by the strategies with column details like date, lot size, trade type (sell or buy), Trading method (Spot, Future, Options) Asset type (Forex, Stock, Crypto), Category (Tech, Agriculture, etc) and any other relevant columns in the table. Store these table of the trades (live trades results) in an suitable table within the system so we can access it the details for machine learning type strategies that the Strategy coder should also be equipped with.